Address Copilot review on PR #12
- swipe_to_act_row: clamp() returns num; make rawTrigger an explicit double
  via toDouble() so _applyResistance's double parameter is unambiguous.
- toast_provider: protect the in-flight queue-activity spinner from
  _maxToasts eviction by removing the oldest non-loading toast instead of
  the oldest overall, so a burst of other toasts can't drop the spinner
  while adds are still running.

+++ lib/widgets/swipe_to_act_row.dart
@@ -191,7 +191,8 @@ class _SwipeToActRowState extends State<SwipeToActRow>
     // Scale the activation distance to the screen so the swipe needs a
     // consistent, reachable finger travel across phone sizes and tablets.
     final rawTrigger = (MediaQuery.sizeOf(context).width * _rawTriggerFraction)
-        .clamp(_rawTriggerMin, _rawTriggerMax);
+        .clamp(_rawTriggerMin, _rawTriggerMax)
+        .toDouble();
     _hapticThreshold = _applyResistance(rawTrigger);
 
     Widget inner;
